requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
anyio==3.6.2
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
einops==0.2.0
fastapi==0.95.1
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.4
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.28.1
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.21.0
lmdb==1.4.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
prefetch-generator==1.0.1
protobuf==4.23.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.0.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
scikit-image==0.20.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.23.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
svgwrite==1.4.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
thop==0.1.1.post2209072238
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
Tree==0.2.4
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.104
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
vietocr==0.3.11
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
